Renaming package to cmdrunner
