+++ pkg/cmdrunner/cmdrunner_test.go
@@ -14,7 +14,7 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-package utils
+package cmdrunner
 
 import "testing"
